requirements.txt
streamlit
transformers
torch
SpeechRecognition
soundfile
pydub
google-genai
